requirements.txt
flask
argon2-cffi
requests
pyjwt==0.4.3
python-dotenv==0.19.2
gunicorn==20.0.4